refactorizar: product details
